+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="0.2.0"></a>
+# 0.2.0 irc-context (2016-08-19)
+
+## Features
+
+- Added `IrcContext` interface providing helper methods, event arguments and information
+- Removed support for the `ACTION` event
+
 <a name="0.1.0"></a>
 # 0.1.0 add-readme (2016-05-28)